backend/requirements.txt
Django
psycopg2-binary
djangorestframework
drf-spectacular
djangorestframework-simplejwt[crypto]
django-cors-headers